+++ src/Products/PloneMeeting/browser/overrides.py
@@ -1750,7 +1750,10 @@ class PMUtils(Utils):
         # do not escape vocabularies used in MeetingConfig
         # especially the SelectableAssemblyMembersVocabulary for which terms
         # contain HTML
-        if not self.request.getURL().endswith('/document-generation') and \
+        if self.request.get('PUBLISHED').__name__ not in (
+            'document-generation',
+            'store-pod-temlate-as-annex-form',
+            'store-items-template-as-annex-batch-action') and \
            not IConfigElement.providedBy(self.context):
             nvalues = [html.escape(val) for val in nvalues]
         if IConfigElement.providedBy(self.context):